SVCPLAN-2149: add gpfs::add_client::interface param
Instead of using the built-in 'ipaddress' fact, which seems to
return the IP address associated with HOSTNAME, allow us to
choose an interface and return its IP address. This allows us
to make sure that when a client is trying to get itself added
to GPFS, it will tell the GPFS master server to use the IP
address associated with its high-speed interface, e.g., IB.

Updates to docs, pdk, etc.